Eliminate compile warnings across local and CI matrix builds
- Add version-portable Math::ThinSVD alias; replaces deprecated
  Eigen::BDCSVD constructor-options form at 6 call sites (AppleClang).
- Fix -Wsign-compare between Eigen signed size() and Index/size_t loop
  counters in PETSc assembly backends, GridFunction, MPI P1, and tests.
- Drop dead locals; mark assert-only reads [[maybe_unused]], keeping the
  assert(ierr == PETSC_SUCCESS) idiom.
- Silence vendored googletest -Wmaybe-uninitialized and mmg
  -Wstringop-overflow per-target (GCC only).
- Suppress Xcode 15 ld duplicate-libraries warning behind
  check_linker_flag.

+++ src/Rodin/PETSc/Variational/GridFunction.h
@@ -345,7 +345,7 @@ namespace Rodin::Variational
           for (const auto& [global, offset] : m_ghosts.right)
           {
             assert(offset >= 0);
-            assert(offset < m_ghosts.left.size());
+            assert(static_cast<size_t>(offset) < m_ghosts.left.size());
             m_ghosts.left[offset] = global;
           }
